requirements.txt
streamlit ==1.32.0
openai==1.14.3
# llama-index==0.10.25
nltk==3.8.1
langchain==0.1.17
langchain-openai==0.1.1
pymupdf==1.24.4
# pypdf==4.1.0
rapidocr-onnxruntime==1.3.21
openpyxl==3.1.2
doc2pdf==0.2.0
docx2pdf==0.1.8
docxpy==0.8.5
fpdf==1.7.2
spacy==3.5.0
scikit-learn==1.4.2
spacy_langdetect==0.1.2
python-dotenv==1.0.1
#en_core_web_md
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0-py3-none-any.whl
faiss-cpu==1.8.0
langchain_experimental==0.0.58
streamlit-pills==0.3.0
tabulate==0.9.0